Move split_at module to its own file
gherrit-pr-id: I797ab16976013a86e8336913109cd4663e021f54
