+++ zerocopy-derive/tests/ui-nightly/struct.stderr
@@ -392,10 +392,10 @@ error[E0277]: the trait bound `SplitAtNotKnownLayout: zerocopy::KnownLayout` is 
               AtomicI16
             and $N others
 note: required by a bound in `SplitAt`
-   --> $WORKSPACE/src/lib.rs
+   --> $WORKSPACE/src/split_at.rs
     |
-    |     pub unsafe trait SplitAt: KnownLayout<PointerMetadata = usize> {
-    |                               ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ required by this bound in `SplitAt`
+    | pub unsafe trait SplitAt: KnownLayout<PointerMetadata = usize> {
+    |                           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ required by this bound in `SplitAt`
     = note: this error originates in the derive macro `SplitAt` (in Nightly builds, run with -Z macro-backtrace for more info)
 
 error[E0277]: the trait bound `u8: SplitAt` is not satisfied
